refactor: change function checkValidator
